fix: Constrain pdm-backend version to fix API change
This package relies on a private method of WheelBuilder to collect
files that will go into the wheel file. There is not a reasonable
public alternative to this function.

We could have alternatively limited pdm-backend to >=2.0.0,<2.02, but
it is preferable to accept future versions of pdm-backend that may have
additional fixes rather than sticking to a much older version.

Fixes #1